refactor: merge availability sidebars into single right panel
Collapse the 3-column availability workspace (left sidebar / main /
right sidebar) into 2 columns by folding the controls panel (mini
calendar, Jupiter toggle, status strip, source list) into the
readiness panel on the right. Removes the standalone "Ask Jupiter"
button/answer card in favor of the single Jupiter toggle+nudge
touchpoint.
