+++ src/assets/locales/en/translation.json
@@ -1603,7 +1603,6 @@
 		},
 		"workspace": {
 			"actions-label": "Availability actions",
-			"ask-jupiter": "Ask Jupiter",
 			"badge-live": "Live",
 			"badge-off": "Off",
 			"badge-on": "On",
@@ -1615,15 +1614,11 @@
 			"checklist-slots-empty": "Create future bookable slots",
 			"checklist-slots-ready": "{{count}} future bookable slots exist",
 			"checklist-title": "Checklist",
-			"close-controls": "Close availability controls",
 			"close-readiness": "Close publish readiness panel",
-			"controls-label": "Availability controls",
-			"jupiter-answer": "I can help review this setup and point you to the next safest change before patients book.",
 			"jupiter-description": "Contextual helper, only when action is needed",
 			"jupiter-label": "Jupiter",
 			"jupiter-suggestion": "Review this week before publishing changes so patients only see safe booking times.",
 			"jupiter-toggle-label": "Toggle Jupiter suggestions",
-			"left-panel-title": "Availability",
 			"month-day-has-slots": "has availability",
 			"month-day-label": "{{date}} {{state}}",
 			"month-day-selected": "selected",
@@ -1631,15 +1626,14 @@
 			"month-nav": "Month navigation",
 			"month-next": "Next month",
 			"month-prev": "Previous month",
-			"open-controls": "Open availability controls",
 			"open-readiness": "Open publish readiness panel",
 			"page-title": "Availability workspace",
+			"publish": "Publish",
 			"publish-changes": "Publish changes",
 			"pwa-note": "PWA rule: preserve deep links, offline-safe shell states, safe-area spacing, and touch-first controls.",
-			"readiness-label": "Publish readiness and selected slot",
+			"readiness-label": "Availability controls and publish readiness",
 			"resolve-readiness": "Review readiness",
-			"right-panel-subtitle": "Review what must be safe before patients book.",
-			"right-panel-title": "Publish readiness",
+			"right-panel-title": "Availability",
 			"skip-calendar": "Skip to calendar",
 			"source-google-desc": "Busy overlay and write-back",
 			"source-google-disconnected-desc": "Connect Google to import busy time",
@@ -1938,10 +1932,11 @@
 			"buffer-label": "Buffer",
 			"legend-available": "Available",
 			"legend-blocked": "Blocked",
-			"legend-booked-google": "Google",
-			"legend-booked-jupiter": "Booked",
+			"legend-booked-google": "Google busy",
+			"legend-booked-jupiter": "Psycron booking",
 			"legend-buffer": "Buffer",
 			"legend-cancelled": "Cancelled",
+			"legend-conflict": "Conflict",
 			"next-week": "Next week",
 			"no-appointments": "No appointments",
 			"page-title": "Availability",
