+++ src/assets/locales/pt/translation.json
@@ -1603,7 +1603,6 @@
 		},
 		"workspace": {
 			"actions-label": "Ações de disponibilidade",
-			"ask-jupiter": "Perguntar ao Jupiter",
 			"badge-live": "Ativo",
 			"badge-off": "Off",
 			"badge-on": "On",
@@ -1615,15 +1614,11 @@
 			"checklist-slots-empty": "Crie horários futuros para agendamento",
 			"checklist-slots-ready": "{{count}} horários futuros disponíveis",
 			"checklist-title": "Checklist",
-			"close-controls": "Fechar controles de disponibilidade",
 			"close-readiness": "Fechar painel de prontidão de publicação",
-			"controls-label": "Controles de disponibilidade",
-			"jupiter-answer": "Posso ajudar a revisar esta configuração e indicar a próxima mudança mais segura antes dos pacientes agendarem.",
 			"jupiter-description": "Ajuda contextual, apenas quando houver ação necessária",
 			"jupiter-label": "Jupiter",
 			"jupiter-suggestion": "Revise esta semana antes de publicar mudanças para que os pacientes vejam apenas horários seguros.",
 			"jupiter-toggle-label": "Alternar sugestões do Jupiter",
-			"left-panel-title": "Disponibilidade",
 			"month-day-has-slots": "tem disponibilidade",
 			"month-day-label": "{{date}} {{state}}",
 			"month-day-selected": "selecionado",
@@ -1631,15 +1626,14 @@
 			"month-nav": "Navegação do mês",
 			"month-next": "Próximo mês",
 			"month-prev": "Mês anterior",
-			"open-controls": "Abrir controles de disponibilidade",
 			"open-readiness": "Abrir painel de prontidão de publicação",
 			"page-title": "Workspace de disponibilidade",
+			"publish": "Publicar",
 			"publish-changes": "Publicar alterações",
 			"pwa-note": "Regra PWA: preserve links diretos, estados offline seguros, espaçamento de área segura e controles pensados para toque.",
-			"readiness-label": "Prontidão de publicação e horário selecionado",
+			"readiness-label": "Controles de disponibilidade e prontidão de publicação",
 			"resolve-readiness": "Revisar prontidão",
-			"right-panel-subtitle": "Revise o que precisa estar seguro antes dos pacientes agendarem.",
-			"right-panel-title": "Prontidão de publicação",
+			"right-panel-title": "Disponibilidade",
 			"skip-calendar": "Ir para o calendário",
 			"source-google-desc": "Sobreposição de ocupado e gravação de volta",
 			"source-google-disconnected-desc": "Conecte o Google para importar horários ocupados",
@@ -1938,10 +1932,11 @@
 			"buffer-label": "Buffer",
 			"legend-available": "Disponível",
 			"legend-blocked": "Bloqueado",
-			"legend-booked-google": "Google",
-			"legend-booked-jupiter": "Agendado",
+			"legend-booked-google": "Google ocupado",
+			"legend-booked-jupiter": "Agendamento Psycron",
 			"legend-buffer": "Buffer",
 			"legend-cancelled": "Cancelado",
+			"legend-conflict": "Conflito",
 			"next-week": "Próxima semana",
 			"no-appointments": "Sem agendamentos",
 			"page-title": "Disponibilidade",
