+++ src/components/availability/AvailabilityLegend/AvailabilityLegend.types.ts
@@ -1,5 +1,6 @@
 export interface AvailabilityLegendItem {
 	borderColor?: string;
+	borderSide?: 'all' | 'left';
 	color: string;
 	label: string;
 	opacity?: number;
