+++ src/pages/availability/workspace/AvailabilityReadinessPanel.tsx
@@ -6,34 +6,95 @@ import {
 	TriangleAlert,
 } from '@psycron/components/icons';
 
+import { AvailabilityMiniCalendar } from './AvailabilityMiniCalendar';
 import {
+	JupiterEyebrow,
+	JupiterIconFrame,
+	JupiterMessage,
+	JupiterNudge,
 	ReadinessActions,
-	ReadinessBody,
 	ReadinessCard,
 	ReadinessCheck,
 	ReadinessCheckIcon,
-	ReadinessJupiterAnswer,
 	ReadinessNote,
 	ReadinessTitle,
+	WorkspaceSwitch,
+	WorkspaceToggleDescription,
+	WorkspaceToggleRow,
+	WorkspaceToggleTitle,
 } from './AvailabilityReadinessPanel.styles';
 import type { AvailabilityReadinessPanelProps } from './AvailabilityReadinessPanel.types';
+import { AvailabilitySourceList } from './AvailabilitySourceList';
+import { AvailabilityStatusStrip } from './AvailabilityStatusStrip';
 
 export const AvailabilityReadinessPanel = ({
-	askJupiterLabel,
+	activeDate,
 	checklistTitle,
 	googleChecklistLabel,
 	hasGoogleConnected,
 	hasSlots,
-	jupiterAnswer,
+	jupiterDescription,
 	jupiterLabel,
+	jupiterSuggestion,
+	jupiterToggleLabel,
 	pwaNote,
 	resolveLabel,
 	slotChecklistLabel,
+	sourceItems,
+	sourcesTitle,
+	statusItems,
+	statusTitle,
 }: AvailabilityReadinessPanelProps) => {
-	const [isJupiterAnswerVisible, setIsJupiterAnswerVisible] = useState(false);
+	const [isJupiterEnabled, setIsJupiterEnabled] = useState(false);
 
 	return (
 		<>
+			<AvailabilityMiniCalendar activeDate={activeDate} />
+			<section
+				aria-labelledby='availability-jupiter-toggle-title'
+				data-testid='availability-jupiter-toggle-section'
+				id='availability-jupiter-toggle-section'
+			>
+				<WorkspaceToggleRow>
+					<div>
+						<WorkspaceToggleTitle id='availability-jupiter-toggle-title'>
+							{jupiterLabel}
+						</WorkspaceToggleTitle>
+						<WorkspaceToggleDescription>
+							{jupiterDescription}
+						</WorkspaceToggleDescription>
+					</div>
+					<WorkspaceSwitch
+						aria-checked={isJupiterEnabled}
+						aria-label={jupiterToggleLabel}
+						data-testid='availability-jupiter-toggle'
+						disabled
+						id='availability-jupiter-toggle'
+						isChecked={isJupiterEnabled}
+						onClick={() => setIsJupiterEnabled((current) => !current)}
+						role='switch'
+						type='button'
+					/>
+				</WorkspaceToggleRow>
+			</section>
+			{isJupiterEnabled ? (
+				<JupiterNudge
+					aria-label={jupiterLabel}
+					aria-live='polite'
+					data-testid='availability-jupiter-nudge'
+					id='availability-jupiter-nudge'
+				>
+					<JupiterIconFrame aria-hidden='true'>
+						<Jupiter />
+					</JupiterIconFrame>
+					<div>
+						<JupiterEyebrow>{jupiterLabel}</JupiterEyebrow>
+						<JupiterMessage>{jupiterSuggestion}</JupiterMessage>
+					</div>
+				</JupiterNudge>
+			) : null}
+			<AvailabilityStatusStrip items={statusItems} title={statusTitle} />
+			<AvailabilitySourceList items={sourceItems} title={sourcesTitle} />
 			<ReadinessCard
 				aria-labelledby='availability-publish-checklist-title'
 				data-testid='availability-publish-checklist'
@@ -42,17 +103,43 @@ export const AvailabilityReadinessPanel = ({
 				<ReadinessTitle id='availability-publish-checklist-title'>
 					{checklistTitle}
 				</ReadinessTitle>
-				<ReadinessCheck>
-					<ReadinessCheckIcon aria-hidden='true' isWarning={!hasGoogleConnected}>
+				<ReadinessCheck
+					id='availability-google-checklist'
+					data-testid='availability-google-checklist'
+				>
+					<ReadinessCheckIcon
+						aria-hidden='true'
+						isWarning={!hasGoogleConnected}
+						id='availability-google-checklist-icon'
+						data-testid='availability-google-checklist-icon'
+					>
 						{hasGoogleConnected ? <CheckSuccess /> : <TriangleAlert />}
 					</ReadinessCheckIcon>
-					<span>{googleChecklistLabel}</span>
+					<span
+						id='availability-google-checklist-label'
+						data-testid='availability-google-checklist-label'
+					>
+						{googleChecklistLabel}
+					</span>
 				</ReadinessCheck>
-				<ReadinessCheck>
-					<ReadinessCheckIcon aria-hidden='true' isWarning={!hasSlots}>
+				<ReadinessCheck
+					id='availability-slots-checklist'
+					data-testid='availability-slots-checklist'
+				>
+					<ReadinessCheckIcon
+						aria-hidden='true'
+						isWarning={!hasSlots}
+						id='availability-slots-checklist-icon'
+						data-testid='availability-slots-checklist-icon'
+					>
 						{hasSlots ? <CheckSuccess /> : <TriangleAlert />}
 					</ReadinessCheckIcon>
-					<span>{slotChecklistLabel}</span>
+					<span
+						id='availability-slots-checklist-label'
+						data-testid='availability-slots-checklist-label'
+					>
+						{slotChecklistLabel}
+					</span>
 				</ReadinessCheck>
 			</ReadinessCard>
 			<ReadinessNote
@@ -62,17 +149,6 @@ export const AvailabilityReadinessPanel = ({
 				{pwaNote}
 			</ReadinessNote>
 			<ReadinessActions>
-				{isJupiterAnswerVisible ? (
-					<ReadinessJupiterAnswer
-						aria-label={jupiterLabel}
-						aria-live='polite'
-						data-testid='availability-jupiter-answer'
-						id='availability-jupiter-answer'
-					>
-						<ReadinessTitle>{jupiterLabel}</ReadinessTitle>
-						<ReadinessBody>{jupiterAnswer}</ReadinessBody>
-					</ReadinessJupiterAnswer>
-				) : null}
 				<Button
 					fullWidth
 					tertiary
@@ -82,20 +158,6 @@ export const AvailabilityReadinessPanel = ({
 				>
 					<TriangleAlert />
 					{resolveLabel}
-				</Button>
-				<Button
-					fullWidth
-					tertiary
-					aria-checked={isJupiterAnswerVisible}
-					aria-controls='availability-jupiter-answer'
-					data-testid='availability-ask-jupiter-button'
-					id='availability-ask-jupiter-button'
-					onClick={() => setIsJupiterAnswerVisible((current) => !current)}
-					role='switch'
-					variant='outlined'
-				>
-					<Jupiter />
-					{askJupiterLabel}
 				</Button>
 			</ReadinessActions>
 		</>
