+++ src/pages/availability/workspace/AvailabilityWorkspacePanel.tsx
@@ -1,15 +1,11 @@
 import { useEffect } from 'react';
-import {
-	ChevronLeft,
-	ChevronRight,
-} from '@psycron/components/icons';
+import { ChevronRight } from '@psycron/components/icons';
 
 import {
 	WorkspacePanelBody,
 	WorkspacePanelClose,
 	WorkspacePanelHeader,
 	WorkspacePanelRoot,
-	WorkspacePanelSubtitle,
 	WorkspacePanelTitle,
 } from './AvailabilityWorkspacePanel.styles';
 import type { AvailabilityWorkspacePanelProps } from './AvailabilityWorkspacePanel.types';
@@ -22,12 +18,13 @@ export const AvailabilityWorkspacePanel = ({
 	isOpen,
 	onClose,
 	panelRef,
-	side,
-	subtitle,
 	testId,
 	title,
 }: AvailabilityWorkspacePanelProps) => {
 	const titleId = `${id}-title`;
+	const headerId = `${id}-header`;
+	const bodyId = `${id}-body`;
+	const closeId = `${id}-close`;
 
 	useEffect(() => {
 		panelRef.current?.toggleAttribute('inert', !isOpen);
@@ -37,31 +34,30 @@ export const AvailabilityWorkspacePanel = ({
 		<WorkspacePanelRoot
 			aria-hidden={!isOpen}
 			aria-label={ariaLabel}
+			aria-labelledby={titleId}
 			data-testid={testId}
 			id={id}
 			isOpen={isOpen}
 			ref={panelRef}
-			side={side}
 		>
-			<WorkspacePanelHeader>
-				<div>
-					<WorkspacePanelTitle id={titleId}>{title}</WorkspacePanelTitle>
-					{subtitle ? (
-						<WorkspacePanelSubtitle>{subtitle}</WorkspacePanelSubtitle>
-					) : null}
-				</div>
+			<WorkspacePanelHeader data-testid={`${testId}-header`} id={headerId}>
 				<WorkspacePanelClose
 					aria-controls={id}
 					aria-label={closeLabel}
 					data-testid={`${testId}-close`}
-					id={`${id}-close`}
+					id={closeId}
 					onClick={onClose}
 					type='button'
 				>
-					{side === 'left' ? <ChevronLeft /> : <ChevronRight />}
+					<ChevronRight aria-hidden='true' />
 				</WorkspacePanelClose>
+				<WorkspacePanelTitle data-testid={`${testId}-title`} id={titleId}>
+					{title}
+				</WorkspacePanelTitle>
 			</WorkspacePanelHeader>
-			<WorkspacePanelBody>{children}</WorkspacePanelBody>
+			<WorkspacePanelBody data-testid={`${testId}-body`} id={bodyId}>
+				{children}
+			</WorkspacePanelBody>
 		</WorkspacePanelRoot>
 	);
 };
