+++ src/pages/availability/workspace/AvailabilityWorkspacePanel.types.ts
@@ -1,7 +1,5 @@
 import type { ReactNode, RefObject } from 'react';
 
-import type { AvailabilityWorkspacePanelSide } from './AvailabilityWorkspaceShell.types';
-
 export interface AvailabilityWorkspacePanelProps {
 	ariaLabel: string;
 	children: ReactNode;
@@ -10,8 +8,6 @@ export interface AvailabilityWorkspacePanelProps {
 	isOpen: boolean;
 	onClose: () => void;
 	panelRef: RefObject<HTMLElement>;
-	side: AvailabilityWorkspacePanelSide;
-	subtitle?: string;
 	testId: string;
 	title: string;
 }
