+++ src/pages/availability/workspace/AvailabilityWorkspaceShell.tsx
@@ -7,7 +7,7 @@ import {
 	useState,
 } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ChevronLeft, ChevronRight } from '@psycron/components/icons';
+import { ChevronLeft } from '@psycron/components/icons';
 import { Loader } from '@psycron/components/loader/Loader';
 import useViewport from '@psycron/hooks/useViewport';
 
@@ -24,7 +24,6 @@ import {
 	WorkspaceScrim,
 } from './AvailabilityWorkspaceShell.styles';
 import type {
-	AvailabilityWorkspacePanelSide,
 	AvailabilityWorkspaceShellHandle,
 	AvailabilityWorkspaceShellProps,
 } from './AvailabilityWorkspaceShell.types';
@@ -49,8 +48,7 @@ export const AvailabilityWorkspaceShell = forwardRef<
 			contentMode = 'calendar',
 			footer,
 			isLoading = false,
-			leftPanel,
-			rightPanel,
+			panel,
 			subtitle,
 			title,
 			viewbar,
@@ -59,57 +57,38 @@ export const AvailabilityWorkspaceShell = forwardRef<
 	) => {
 		const { t } = useTranslation();
 		const { isSmallerThanTablet } = useViewport();
-		const [openPanel, setOpenPanel] =
-			useState<AvailabilityWorkspacePanelSide | null>(null);
-		const leftPanelRef = useRef<HTMLElement>(null);
-		const rightPanelRef = useRef<HTMLElement>(null);
-		const leftToggleRef = useRef<HTMLButtonElement>(null);
-		const rightToggleRef = useRef<HTMLButtonElement>(null);
+		const [isPanelOpen, setIsPanelOpen] = useState(false);
+		const panelRef = useRef<HTMLElement>(null);
+		const toggleRef = useRef<HTMLButtonElement>(null);
 
-		const isLeftOpen = openPanel === 'left';
-		const isRightOpen = openPanel === 'right';
-		const isAnyPanelOpen = openPanel !== null;
-
-		const getPanelElement = useCallback(
-			(side: AvailabilityWorkspacePanelSide): HTMLElement | null =>
-				side === 'left' ? leftPanelRef.current : rightPanelRef.current,
-			[]
-		);
-
-		const focusToggle = useCallback(
-			(side: AvailabilityWorkspacePanelSide): void => {
-				const toggle =
-					side === 'left' ? leftToggleRef.current : rightToggleRef.current;
-				toggle?.focus();
-			},
-			[]
-		);
+		const focusToggle = useCallback((): void => {
+			toggleRef.current?.focus();
+		}, []);
 
 		const closePanel = useCallback(
 			(shouldReturnFocus = true): void => {
-				const side = openPanel;
-				setOpenPanel(null);
-				if (side && shouldReturnFocus) {
-					window.requestAnimationFrame(() => focusToggle(side));
+				setIsPanelOpen(false);
+				if (shouldReturnFocus) {
+					window.requestAnimationFrame(() => focusToggle());
 				}
 			},
-			[focusToggle, openPanel]
+			[focusToggle]
 		);
 
-		const togglePanel = (side: AvailabilityWorkspacePanelSide): void => {
-			setOpenPanel((current) => (current === side ? null : side));
+		const togglePanel = (): void => {
+			setIsPanelOpen((current) => !current);
 		};
 
 		useImperativeHandle(
 			ref,
 			() => ({
-				openRightPanel: () => setOpenPanel('right'),
+				openPanel: () => setIsPanelOpen(true),
 			}),
 			[]
 		);
 
 		useEffect(() => {
-			if (!openPanel) return;
+			if (!isPanelOpen) return;
 
 			const handleKeyDown = (event: KeyboardEvent): void => {
 				if (event.key === 'Escape') {
@@ -119,11 +98,11 @@ export const AvailabilityWorkspaceShell = forwardRef<
 
 				if (!isSmallerThanTablet || event.key !== 'Tab') return;
 
-				const panel = getPanelElement(openPanel);
-				if (!panel) return;
+				const panelElement = panelRef.current;
+				if (!panelElement) return;
 
 				const focusableElements = Array.from(
-					panel.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)
+					panelElement.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)
 				).filter((element) => !element.hasAttribute('disabled'));
 
 				if (focusableElements.length === 0) return;
@@ -144,22 +123,22 @@ export const AvailabilityWorkspaceShell = forwardRef<
 
 			document.addEventListener('keydown', handleKeyDown);
 			return () => document.removeEventListener('keydown', handleKeyDown);
-		}, [closePanel, getPanelElement, isSmallerThanTablet, openPanel]);
+		}, [closePanel, isPanelOpen, isSmallerThanTablet]);
 
 		useEffect(() => {
-			if (!openPanel || !isSmallerThanTablet) return;
+			if (!isPanelOpen || !isSmallerThanTablet) return;
 
-			const panel = getPanelElement(openPanel);
-			if (!panel) return;
+			const panelElement = panelRef.current;
+			if (!panelElement) return;
 
 			const frameId = window.requestAnimationFrame(() => {
 				const firstFocusable =
-					panel.querySelector<HTMLElement>(FOCUSABLE_SELECTOR);
+					panelElement.querySelector<HTMLElement>(FOCUSABLE_SELECTOR);
 				firstFocusable?.focus();
 			});
 
 			return () => window.cancelAnimationFrame(frameId);
-		}, [getPanelElement, isSmallerThanTablet, openPanel]);
+		}, [isPanelOpen, isSmallerThanTablet]);
 
 		return (
 			<WorkspaceRoot
@@ -174,64 +153,32 @@ export const AvailabilityWorkspaceShell = forwardRef<
 					data-testid='availability-workspace'
 					id='availability-workspace'
 				>
-					<WorkspaceEdgeToggle
-						aria-controls='availability-left-sidebar'
-						aria-expanded={isLeftOpen}
-						aria-label={t(
-							isLeftOpen
-								? 'availability.workspace.close-controls'
-								: 'availability.workspace.open-controls'
-						)}
-						data-testid='availability-left-sidebar-toggle'
-						id='availability-left-sidebar-toggle'
-						isAnyPanelOpen={isAnyPanelOpen}
-						onClick={() => togglePanel('left')}
-						panelSide='left'
-						ref={leftToggleRef}
-						type='button'
-					>
-						<ChevronRight />
-					</WorkspaceEdgeToggle>
 					<WorkspaceScrim
 						aria-hidden='true'
 						data-testid='availability-drawer-scrim'
 						id='availability-drawer-scrim'
-						isVisible={isAnyPanelOpen}
+						isVisible={isPanelOpen}
 						onClick={() => closePanel()}
 						tabIndex={-1}
 						type='button'
 					/>
 					<WorkspaceEdgeToggle
 						aria-controls='availability-right-sidebar'
-						aria-expanded={isRightOpen}
+						aria-expanded={isPanelOpen}
 						aria-label={t(
-							isRightOpen
+							isPanelOpen
 								? 'availability.workspace.close-readiness'
 								: 'availability.workspace.open-readiness'
 						)}
 						data-testid='availability-right-sidebar-toggle'
 						id='availability-right-sidebar-toggle'
-						isAnyPanelOpen={isAnyPanelOpen}
-						onClick={() => togglePanel('right')}
-						panelSide='right'
-						ref={rightToggleRef}
+						isOwnPanelOpen={isPanelOpen}
+						onClick={togglePanel}
+						ref={toggleRef}
 						type='button'
 					>
-						<ChevronLeft />
+						<ChevronLeft aria-hidden='true' />
 					</WorkspaceEdgeToggle>
-					<AvailabilityWorkspacePanel
-						ariaLabel={t('availability.workspace.controls-label')}
-						closeLabel={t('availability.workspace.close-controls')}
-						id='availability-left-sidebar'
-						isOpen={isLeftOpen}
-						panelRef={leftPanelRef}
-						side='left'
-						testId='availability-left-sidebar'
-						title={t('availability.workspace.left-panel-title')}
-						onClose={() => closePanel()}
-					>
-						{leftPanel}
-					</AvailabilityWorkspacePanel>
 					<WorkspaceMain
 						aria-labelledby='availability-page-title'
 						data-testid='availability-main'
@@ -265,15 +212,13 @@ export const AvailabilityWorkspaceShell = forwardRef<
 						ariaLabel={t('availability.workspace.readiness-label')}
 						closeLabel={t('availability.workspace.close-readiness')}
 						id='availability-right-sidebar'
-						isOpen={isRightOpen}
-						panelRef={rightPanelRef}
-						side='right'
-						subtitle={t('availability.workspace.right-panel-subtitle')}
+						isOpen={isPanelOpen}
+						panelRef={panelRef}
 						testId='availability-right-sidebar'
 						title={t('availability.workspace.right-panel-title')}
 						onClose={() => closePanel()}
 					>
-						{rightPanel}
+						{panel}
 					</AvailabilityWorkspacePanel>
 				</WorkspaceFrame>
 			</WorkspaceRoot>
